refactor(geoadmin): share the request budget and the tile store
A second feature is about to call api3.geo.admin.ch, and the service's fair-use
limit is 40 req/min for the whole client. Two features each running their own
30 req/min limiter would issue 60 and get throttled, so the limiter cannot be
duplicated: it is a correctness constraint, not a DRY preference.

Moves the transport and the rate limiter to src/geoadmin/http.ts, and the
IndexedDB tile cache to src/geoadmin/idb-store.ts, now generic over its entry
type and parameterized by database name, store name and version. The checker
keeps its own database (IndexedDB versions the database, not the object store,
and its upgrade handler drops the store), and becomes a three-line subclass that
pins that identity.

None of the checker's own classes or CSS selectors are renamed; only one import
line moves in its test, and no assertion changes.
