+++ src/geoadmin/types.ts
@@ -1,0 +1,2 @@
+/** [minLon, minLat, maxLon, maxLat] in WGS84. */
+export type Bbox = [number, number, number, number];
