+++ src/street-name-checker/geoadmin/types.ts
@@ -24,5 +24,4 @@ export interface OfficialStreet {
   lines: number[][][] | null;
 }
 
-/** [minLon, minLat, maxLon, maxLat] in WGS84. */
-export type Bbox = [number, number, number, number];
+export type { Bbox } from "../../geoadmin/types";
